feat(simulation+model options):
Simulation and model file initialization parameters and attributes are now dynamically generated to include the namefile options found in the dfn file.
