+++ flopy/mf6/modflow/mfgwt.py
@@ -1,6 +1,6 @@
 # DO NOT MODIFY THIS FILE DIRECTLY.  THIS FILE MUST BE CREATED BY
 # mf6/utils/createpackages.py
-# FILE created on January 27, 2023 18:36:16 UTC
+# FILE created on June 22, 2023 21:13:41 UTC
 from .. import mfmodel
 from ..data.mfdatautil import ArrayTemplateGenerator, ListTemplateGenerator
 
@@ -84,7 +84,6 @@ class ModflowGwt(mfmodel.MFModel):
         print_input=None,
         print_flows=None,
         save_flows=None,
-        packages=None,
         **kwargs,
     ):
         super().__init__(
@@ -102,7 +101,11 @@ class ModflowGwt(mfmodel.MFModel):
         self.name_file.print_input.set_data(print_input)
         self.name_file.print_flows.set_data(print_flows)
         self.name_file.save_flows.set_data(save_flows)
-        self.name_file.packages.set_data(packages)
+
+        self.list = self.name_file.list
+        self.print_input = self.name_file.print_input
+        self.print_flows = self.name_file.print_flows
+        self.save_flows = self.name_file.save_flows
 
     @classmethod
     def load(
@@ -118,6 +121,7 @@ class ModflowGwt(mfmodel.MFModel):
         load_only=None,
     ):
         return mfmodel.MFModel.load_base(
+            cls,
             simulation,
             structure,
             modelname,
